fix: endpoint will not be select if disabled
